+++ cloud/backend/meta-merger/meta-merger.ts
@@ -1,11 +1,10 @@
-import { ps } from "@fireproof/core";
+import { ps, rt } from "@fireproof/core";
 import { MetaByTenantLedgerSql } from "./meta-by-tenant-ledger.js";
 import { MetaSendSql } from "./meta-send.js";
 import { TenantLedgerSql } from "./tenant-ledger.js";
 import { TenantSql } from "./tenant.js";
 import { SQLDatabase } from "./abstract-sql.js";
 import { Logger } from "@adviser/cement";
-import { V2SerializedMetaKey } from "../../src/runtime/meta-key-hack.js";
 import { KeyByTenantLedgerSql } from "./key-by-tenant-ledger.js";
 
 type TenantLedger = ps.cloud.TenantLedger;
@@ -19,7 +18,7 @@ export interface Connection {
 export interface MetaMerge {
   // readonly logger Logger;
   readonly connection: Connection;
-  readonly meta: V2SerializedMetaKey;
+  readonly meta: rt.V2SerializedMetaKey;
   readonly now?: Date;
 }
 
@@ -132,7 +131,7 @@ export class MetaMerger {
     }
   }
 
-  async metaToSend(sink: Connection, now = new Date()): Promise<V2SerializedMetaKey> {
+  async metaToSend(sink: Connection, now = new Date()): Promise<rt.V2SerializedMetaKey> {
     const bySink = toByConnection(sink);
     const rows = await this.sql.metaSend.selectToAddSend({ ...bySink, now });
     // console.log("metaToSend", bySink, rows);
