+++ cloud/backend/pre-signed-url.ts
