+++ vitest.cloud.config.ts
@@ -15,7 +15,7 @@ export default defineConfig({
       "tests/gateway/file",
       "tests/blockstore/keyed-crypto-indexeddb-file.test.ts",
     ],
-    include: ["tests/**/*test.?(c|m)[jt]s?(x)", "cloud-backend/**/*test.?(c|m)[jt]s?(x)"],
+    include: ["tests/**/*test.?(c|m)[jt]s?(x)", "cloud/**/*test.?(c|m)[jt]s?(x)"],
     coverage: {
       exclude: ["**/smoke/**", "**/scripts/**", "**/examples/**"],
     },
